Add documentation for the user flag
This flag was recently added and allows the system administrator to
change the user and group that will run the daemon. The daemon can only
switch to another user if it is itself run with root privileges
initially. If you have a script to start the server under another user,
this is unnecessary.